fix: resolve all compiler warnings in C++ codebase
Fixed all compiler warnings across the plugin codebase:

**Unused Parameter Warnings**:
- Added comment-style suppressions for intentionally unused parameters
- Fixed in OSNotification, ErrorTracking, WebSocketClient, StreamChatClient,
  AudioCapture, MIDICapture, Microphone, ChordSequenceDetector

**Sign Conversion Warnings**:
- Added explicit static_cast<> conversions for int->size_t conversions
- Fixed in HttpErrorHandler, PerformanceMonitor, WebSocketClient,
  StreamChatClient, ErrorTracking, LogSink, OperationalTransform

**Shadow Warnings**:
- Renamed local variables to avoid shadowing class members
- Fixed in StreamChatClient (renamed token/apiKey/userId)
- Fixed in PerformanceMonitor (renamed name to metricName)

**Returning Reference to Temporary**:
- Changed DraftStore::getDrafts() return type from const ref to value

**Nullable to Non-nullable Conversion**:
- Added null checks in OSNotification_mac.mm before NSString assignments

All fixes maintain C++26 compatibility and don't change functionality.

+++ plugin/src/util/error/ErrorTracking.cpp
@@ -182,7 +182,7 @@ ErrorTracker::ErrorStats ErrorTracker::getStatistics() const
     std::sort(sortedErrors.begin(), sortedErrors.end(),
               [](const auto& a, const auto& b) { return a.second > b.second; });
 
-    for (int i = 0; i < std::min(10, static_cast<int>(sortedErrors.size())); ++i)
+    for (size_t i = 0; i < std::min(static_cast<size_t>(10), sortedErrors.size()); ++i)
     {
         stats.topErrors[sortedErrors[i].first] = sortedErrors[i].second;
     }
